Refactor shared dbt utilities
