+++ dagster_pipeline.py
@@ -1,45 +1,10 @@
 import importlib
 import os
-import subprocess
-import sys
-from pathlib import Path
-
 from s3_utils import download_seeds
-
-import yaml
+from utils import DBT_DIR, _run_dbt, load_config, active_models
 from dagster import Definitions, ScheduleDefinition, job, op, Field, Noneable
 
-DBT_DIR = Path(__file__).parent / "dbt"
-CONFIG_FILE = Path(__file__).parent / "pipeline_config.yml"
-
 os.environ.setdefault("DBT_PROFILES_DIR", str(DBT_DIR))
-
-
-def _run_dbt(args: list[str]) -> None:
-    """Execute a dbt command with a fallback to ``python -m dbt``.
-
-    Raises a ``RuntimeError`` with captured output if the command fails or
-    when dbt is not installed."""
-    commands = [["dbt", *args], [sys.executable, "-m", "dbt", *args]]
-    last_error: Exception | None = None
-    for cmd in commands:
-        try:
-            result = subprocess.run(
-                cmd,
-                cwd=DBT_DIR,
-                capture_output=True,
-                text=True,
-            )
-        except FileNotFoundError as exc:
-            last_error = exc
-            continue
-        if result.returncode == 0:
-            return
-        last_error = RuntimeError(
-            f"{' '.join(cmd)} failed with code {result.returncode}\n{result.stdout}\n{result.stderr}"
-        )
-    if last_error:
-        raise last_error
 
 
 @op(
@@ -91,14 +56,6 @@ def pipeline_job() -> None:
     run_dbt_pipeline(fetch_data())
 
 
-def load_config() -> dict:
-    with open(CONFIG_FILE) as f:
-        return yaml.safe_load(f) or {}
-
-
-def active_models(cfg: dict) -> set[str]:
-    models = cfg.get("models", [])
-    return {m["name"] for m in models if m.get("active")}
 
 
 def parse_cron(expr: str) -> str:
