+++ run_pipeline.py
@@ -1,51 +1,9 @@
 """Run a single pipeline execution without Dagster."""
-
 import argparse
 import importlib
-import subprocess
-import sys
-from pathlib import Path
 
-import yaml
 from s3_utils import download_seeds
-
-DBT_DIR = Path(__file__).parent / "dbt"
-CONFIG_FILE = Path(__file__).parent / "pipeline_config.yml"
-
-
-def _run_dbt(args: list[str]) -> None:
-    """Execute a dbt command with a fallback to ``python -m dbt``.
-
-    Raises ``RuntimeError`` if the command fails or dbt is not installed."""
-    commands = [["dbt", *args], [sys.executable, "-m", "dbt", *args]]
-    last_error: Exception | None = None
-    for cmd in commands:
-        try:
-            result = subprocess.run(
-                cmd,
-                cwd=DBT_DIR,
-                capture_output=True,
-                text=True,
-            )
-        except FileNotFoundError as exc:
-            last_error = exc
-            continue
-        if result.returncode == 0:
-            return
-        last_error = RuntimeError(
-            f"{' '.join(cmd)} failed with code {result.returncode}\n{result.stdout}\n{result.stderr}"
-        )
-    if last_error:
-        raise last_error
-
-
-def load_config() -> dict:
-    with open(CONFIG_FILE) as f:
-        return yaml.safe_load(f) or {}
-
-
-def active_models(cfg: dict) -> list[str]:
-    return [m.get("name") for m in cfg.get("models", []) if m.get("active")]
+from utils import DBT_DIR, _run_dbt, load_config, active_models
 
 
 def fetch(fetcher: str) -> None:
@@ -83,7 +41,7 @@ def main() -> None:
         else:
             parser.error("No fetcher specified and no sources in config")
 
-    models = args.models if args.models is not None else active_models(cfg)
+    models = args.models if args.models is not None else list(active_models(cfg))
 
     fetch(fetcher)
     run_dbt(models)
